properly fix race condition on tagging
Mistaking wait(2) for the same as Bash's wait (wait for _all_ children
before returning instead of waiting for _any_ child) was the initial
reason for a race-condition that led to tagging before encoding was
finished.

This is now properly fixed  -- no stupid call to `sync` anymore.
An unneeded prefork was removed as well.